Streamline exception classes and names
Expected errors (those errors that are caused by the environment or
input data and not bugs in the code) are now all subclasses of
MachineryError. Those exceptions will be handled specially, e.g. by not
showing a backtrace.

This commit also harmonizes the names of the exception class which were
not very consistent.