Handle unexpected errors appropriately
Excpetions which aren't treated specially (like MachineryException or
Cheetah::ExecutionFailed) are considered bugs in our code so we log the
error and the backtrace and show a meaningful message to the user. That
message explains how to file a bug report for machinery and contains all
relevant information about the error including the backtrace.