Do not continue inspecting the system in case of unexpected errors
When an Inspector fails with an "expected error" (inheriting from
MachineryError) we collect those errors and show them to the user after
the full inspection is done.

In case of unexpected errors we rather want to bail out immediately
though because it's likely that the whole inspection result will be
useless.