Fix "analyze -o config-file-diffs" on SLES12 systems
Until now we called zypper so that temporary directories were used
instead of the system ones e.g. for the package cache and other things.
Unfortunately zypper does not allow to override the services
directory which is a problem on systems that have services installed.

In order to solve that conflict zypper is now called in a complete
chroot environment, making it completely independent of the system
zypper environment.