Add SystemDescription method to get Os instance
The SystemDescription#os_object has the minimal implementation for
now. This needs later refactoring to move the name comparisons to
the specific Os subclasses.

The method is names os_object for now to not conflict with the
os scope which is accessible via the os method. This should probably
be merged somehow in the future.