Use native Os objects when checking for platform
The test used an invalid OS name, which is not legal in the native
objects, so we now use a valid OS name, which is not supported for
building.

This could be refactored further by putting the Os specific code
into the Os classes. Then the case statement would not be needed
anymore. Drawback would be that the build specific code is then
distributed over more than one class.