Shellescape directory names in order to not choke on special chars
The "UnmanagedFilesInspector" has some shell code which is not run through
the proper means like "System#run_command" or "Cheetah#run_command" but
is given to bash on STDIN instead. That means that we have to manually escape
the variables that are expanded into that code, otherwise it chokes on
spaces and ticks etc.

Fixes #21