Data format cleanup: Change error reporting for changed_managed_files
Introduce a "status" attribute for files in changed_managed_files scope.
Its value is either "changed" (in which case the changes are described
in the "changes" attribute) or "error" (in which case the error
encountered when detecting changes is described in the "error_message"
attribute).

This replaces previous "error" attribute. The advantage is that readers
don't have to rely on its presence/absence to determine file state.