Implement missing Machinery::Array#hash method
Various Array operators such as "-" and "&" use #eql? and #hash to
compare array elements. Because #hash was undefined for
Machinery::Array, during comparison of nested Machinery::Array
structures it could have happened that inner Machinery::Array instances
that were equal were considered different (becasuse the default #hash
implementation returns the same result only for the same objects
reference-wise). The comparison returned wrong result as a consequence.

This commit implements the missing Machinery::Array#hash method as a
delegator to the inner @elements array. Because Array (the class of
@elements) implements #hash sanely (structurally), this solves the
problem and fixes the comparison.

Part of a fix of #23.