Change Machinery::Scope#compare_with implementation
Machinery::Scope#compare_with used a simple hard-coded implementation,
which didn't return correct results when the payload was an array (which
is actually the case for most scopes).

This commit changes the implementation to delegate comparison to the
payload. This required implementing Machinery::Object#compare_with
method, which was missing.

Part of a fix of #23.