Fix rendering of array comparison
When comparing two scopes containing arrays of items, nothing was
printed by Renderer#render_comparison in case one of the arrays
contained all the items as the other one plus some additional ones.

This commit fixes the code in Renderer#render_comparison to print the
additional items. The commit also restores a test for this condition,
which was wrongly re-purposed to test something completely different.

Part of a fix of #23.