Unify scope parsing and checking for their existance
I have added the Cli method process_scope_option to parse the scopes and
check for their existance.
This does also allow the consolidation of the scope error handling.

One requirement for the scope check to work is that each inspector has a
renderer with the same name an vice versa.
This is at the moment the case but if this changes in the future we must
pass along if the scope list is relevant for the renderer or the
inspector.