Don't save version metadata for descriptions without format version
Saving version metadata for old descriptions without format version
resulted in JSON like this:

  {
    "meta": {
      "format_version": null
    }
    ...
  }

I'd rather like "format_version" not to be present in this case (so that
we can say it must always be an integer if it is there).

Saving old descriptions happens only in the tests, so this is quite an
edge case, but I think that cleaner and more consistent description
saving code can't hurt.