Replace string-scrub gem with our own implementation
In UnmanagedFilesInspector, we need to clean invalid UTF-8 byte
sequences. This would be a job for String#scrub in Ruby 2.1, but because
we have to be Ruby 2.0 compatible, we couldn't use it directly and we
had to resort to the string-scrub gem.

This gem had to be packaged for both openSUSE and SLES so that our gem
bundling mecahnism works. However packaging for SLES was met with some
resistance as Ruby 2.1 is available there and the gem isn't strictly
needed. Instead of fighting we decided to just implement the scrubbing
ourselves, which turned out to be relatively simple in our case (only
UTF-8).