Add "bundle check || sudo bundle install" into prophet.rb
These commands ensure that pull requests which add new gems to the
dependencies won't fail when Prophet executes the tests.

I used "sudo" in case of "bundle install" to avoid possible password
prompts. It's not very clean, but we use it already in scripts on
Jenkins without any negative consequences.