Add tests for rendering comparisons
This adds a test for rendering comparisons of packages as an example for
comparing list and a test for rendering comparisons of the os as an
example for comparing objects.

The packages comparison fails because of issue #23. This will be fixed in
the next two commits.

Based on code by Cornelius Schumacher from #44.

Part of a fix of #23.