Implement validation of existence of extracted config files
This patch implements the first step of the file validation. It
validates that for each file listed in the manifest of an extracted
system description an actual file exists in the corresponding sub
directory.

This patch only implements the validation for the config-files scope.
Other scopes will be implemented in subsequent patches.

The validation for the scope is currently implemented in the
SystemDescription class. It eventually should be moved to the more
specific scope class to reduce the dependency of the SystemDescription
on scope specific behavior. This will likely require some more
refactoring of how file stores are handled.

The tests are in a seperate file for now as there will be more tests
for other scopes and other types of validation. This way they don't
clutter the main SystemDescription tests. This also will make it easier
if the validation code is moved to other classes.