Revert "Convert scope name in schema error message"
It doesn't make sense to transform scope names from underscore-separated
form to dash-separated form in JSON validation error messages. The scope
names should be exactly the same as in the JSON file to prevent any
confusion, make it possible to grep in manifest.json directly from the
error message, etc.

Beside this, the Machinery::UI.internal_scope_list_to_string method
wasn't the right tool for the job. It is a UI-level method, not a
utility one, and it is supposed to convert a list of scopes, not one.

This reverts commit fcd017936b7ef5da8a9f7cc805e49ba66449e4b8.

Conflicts:
	lib/system_description.rb