Schemas: Change structure of the "unmanaged_files" schema
Change structure of the "unmanaged_files" schema to eliminate
duplication and make the structure more logical.

Also replace the the '"additionalProperties": false' hack in the
"not_extracted" definition by the "not" keyword. The hack was there to
make the definition mutually-exclusive with "extracted_*" definitions.
(without this, the toplevel "oneOf" keyword wouldn't work). However, it
prevented extending JSON objects representing non-extracted unmanaged
files, which was wrong per our data model design (which says the user
should be able to extend any object with custom properties).