Refactor validation of presence of extracted config-files
The actual validation that all expected files exist is moved from
SystemDescriptionStore to SystemDescription. That way the store class
doesn't have to have knowledge about scope internals like the deleted
flag.

Eventually the validation will be moved to the specialized scope
classes, but for now I think it's cleaner to have it in the
SystemDescription rather than the store class.