Extract JSON validation code into SystemDescriptionValidator
The validation code in SystemDescription grew enough to warrant a new
class -- SystemDescriptionValidator. In this commit, all JSON validation
code is moved there and tests are adapted.

Note I didn't attempt to do any cleanups in the code nor in the tests
(beside removing some now useless "describe" blocks). This will come
later. As a result, this is mostly a mechanical change.