Use tmp directory for extracting unmanaged files
During inspection unmanaged files are extracted to the file store
"unmanaged_files.tmp".
The directory gets renamed as soon as the inspection is finished.

This allows debugging after canceling the inspection and also prevents
overwriting of working data.

If unmanaged_files were extracted before the file store is kept until
the new inspection is finished to keep the system description state
consistent.

Fixes #160