Fix file store rename for integration test
Renaming of the file store failed during the integretation test because
of permission issues besides working locally and also when the
inspection was started manually in Pennyworth.

It is related to the vagrant environment because the commands were run
with su without the "-l" option. I have fixed the issue in
Pennyworth but using absolute paths in rename is still fine.