Create a SystemDescriptionFactory module for the RSpec tests
The module is available in all specs and allows to easily create
SystemDescriptions either in-memory or in a temporary
SystemDescriptionStore on disk. It also allows for
storing raw descriptions which aren't compatible with the code anymore.

Example for creating a simple description in-memory:

  create_test_description(json: "{}", name: "name1")

More complex example that uses the temporary store:

  describe Foo
    initialize_system_description_factory_store

    it "does something to the description" do
      description = create_test_description(
        store_on_disk: true,
        json: test_json
      )
    end
  end