Exclude remote file systems from inspection of unmanaged files
Remote file systems are all file systems which are mounted either via NFS or
CIFS, directly or via the auto mounter. These file systems are not extracted
by the unmanaged-files inspector. The mount points of remote file systems are
stored as entry of type 'remote-dir' in the unmanaged-files scope in the
system description.