Transform include_scope matcher into include_file_scop matcher
The include_scope matcher seemed to be applicable to both
Machinery::Arrays and Machinery::Objects, but it actually only worked
for Machinery::Arrays. It was also only used to match file scopes which
were changed to Machinery::Objects recently so it broke there as well.

This commit fixes the matcher for the file scopes and makes it clear
that it only works for file scopes from the name.