Allow for filling the test descriptions with scope content
The system description factory now has two new options. The scopes given
in the "scopes" options will be added to the system description with
some example content. The scopes in the "extracted_scopes" option will
additionally be marked as "extracted" and the according files will be
written to disk if "store_on_disk" is true.