Implement HTML view on system descriptions
This commit adds two new commands to Machinery for creating and showing
an HTML view of system descriptions:

1. machinery generate-html DESCRIPTION

Generates the according HTML page (index.html) and puts it in the system
description directory.

2. machinery show DESCRIPTION --html

generates the HTML page and opens it in the browser using "xdg-open".

With this commit we also dynamically generate scope documentation for
the man page and the HTML exports from one source (markdown files in
plugins/docs/)