Basic workflow example in man page
User struggle with understanding the concept of Machinery, when
they don't take the time to read through all the man page. A
quick introduction with examples would help with that.

Since there are already enough descriptions about machinery and
its concept in the man page I just provided examples as a
quickstart guide.