Store class level Os info in class instead of object
Move the information about the canonical operating system name, the
buildable systems and whether a system can run machinery to the class.
This simplifies the code by getting rid of the member variables which
were used to store the info per object and creates a better model,
because this information is the same for all objects of a class by
definition.

The canonical name is the expected name used to identify an operating
system. The method has been renamed to avoid confusion with the
name stored in the os scope, which is the data actually read from
disk.