Merge Os and OsScope
Both classes have the same purpose, describing an operating system.
Os contains the class level information, and OsScope the object
level information. This naturally fits into the same class.

This is the first step of the merge. The creation of the Scope objects
and determination of the scope name which currently is special-cased
needs to be generalized. SystemDescription.os_object needs to be
replaced by the os method.