Change args order when creating SystemDescription
Make the SystemDescriptionStore object the second argument of the
constructor and move the JSON object to the third argument. This
reflects that all descriptions need some sort of storage, but only
some are constructed with an initial JSON object. The next step
will be to make the store parameter non-optional.