Make store mandatory for SystemDescription
The constructor of SystemDescription now takes a mandatory store
argument, so that all system descriptions have an explicit way of
storing the description.

For the cases where we only need a transient description which is
not supposed to be saved, there is the SystemDescriptionStoreMemory
class which provides transient storage and fails when it is
attempted to store data to disk.