Prevent class overloading by stubbing the migration classes
I ran into issues that my migrate2to3 failed if I ran it with all unit
tests while it worked fine when run alone.
The reason for this was that in migration_spec the class Migrate2To3 was
already specified so it messed the test up.

To prevent this the class was stubbed in the migration spec.