Add method to get ScopeFileStore
Add method to get a ScopeFileStore object for a given scope from
the SystemDescription. This can then be used to operate on the
directory represented by this object.

The code accessing the file store can then be refactored to use
this object instead of the old methods in SystemDescription and
SystemDescriptionStore.