Fix non working network interfaces on autoyast installed sle products.
We did not filter the 70-persistent-net.rules udev rule and therefor the
installed system was lacking a working network environment. This fix filters
the udev rule from being exported to autoinst.xml.
Background: SLE product use a persistent network device setup which binds a
interface configuration to a certain mac address. However, in many cases an
installation is performed on a different system than the inspected one with
a different mac address. In these cases the installation would finish
without working network interfaces if we would not filter the udev rule for
this behavior.