Prevent missing_scope comparison messages if both descriptions lack them
In doesn't make much sense in case of a comparison to inform about
missing scopes if both don't have them.
It also looks weird to inform about missing_scopes in case of an
identical comparison.
With this commit the missing scope message is only shown if it is
missing from only on scope.

Fixes #498