Remove alternating files during test host provision
Remove certain files during provision of the test host which have
alternating file names with dates, or changing log names and so on
to make the unmanaged-file tests more stable.

I think the yum history information could be useful so it is not
filtered anymore during inspection but removed during provision
instead.