Raise error when os cannot be analyzed
With this patch the `analyze` command checks the description for
the operating system and raises an error, when the OS cannot be
analyzed.

To have a nicer way of printing the name of the OS in the error
message this patch also introduces a `display_name` attribute on
the Os class.

If an unknown operating system is found, the name is not prefixed
with "Unknown" anymore. If Machinery knows the OS or not can be
handled through the OsUnknown class, no need to distort the name
for the user.