Make the SystemValidator work on raw data without using the data model
The SystemValidator should be able to validate SystemDescriptions which
use the current format version, but also descriptions which use older
formats. That means that it can't rely on the data model which might no
longer be compatible.

It also means that it has to have different code paths for validating
the file structure in the different format versions because that
validation is done in code, not using schema files like the JSON
validation does.