Add SystemDescription.load 'skip_validation' option
To allow list_task to be much faster an option was added to
SystemDescription.load to prevent the validation which improves
the loading time by the factor 10 on the test system.

The unit tests are mostly implementation tests but we need our system
description factory to be able to create broken system descriptions
first to tests this with real data.