Move unmanaged files inspector to new filter class
This is a first step, which exactly replicates the behavior. It
needs additional cleanup. Nnot everything which is tested makes
sense. The filter definitions and tests can be simplified.

The cleanup is a bit tricky, because the unit tests don't cover
the filtering very well. The integration tests run for a long time
and they also might not cover every situation.