Give system and description to the Inspector constructor already
Since the summary is now generated separately and not as a part of the
"inspect" call anymore it is cleaner to keep the description as an
instance variable instead of giving it to "inspect" and "summary"
separately.

Also the system is rather an attribute of the Inspector
instance than a parameter for the actual "inspect" call.