Also transform payload into rich data model classes in #new
Until now we only used the data model for the objects when loading it
from json using the #from_json methods of Machinery::Array and
Machinery::Object, but not when initializing new objects with #new.

That problem was not discovered until now because #new was only used
during inspection and there the system description was only written as a
json file where the data model was not needed. The problem showed up
only now that we started filtering the new system descriptions.