Refactor InspectTask to clear up the different filter objects
The "filter" object is the Filter specified by the user and the default
filter definitions.

"effective_filter" is the Filter that is being applied to the description
during the inspection. This filter only contains the filters for the scopes
that are inspected in order to not overwrite or erase the filters of other
scopes that might already exist in the description.