Do not use rpm verification scripts
This commit adds the `--noscripts` argument to rpm verification calls, which makes verification scripts not to be executed. The output of verification scripts is not standardized and therefor it cannot be parsed by Machinery. That's why we don't want verification scripts to be executed. The version of rpm initially shipped with SLES11SP3 doesn't support `--noscripts`. A workaround is used in this case.