Run integration tests controlled by a jeos file.
- Main purpose
The inspect, build, analyze, export and deploy test_examples are now run
based on entries in alfred/documents/source/integration_tests.yml.

- The system machinery is running on gets defined via test description
We retrieve the name of the system machinery is executed on from the
test description in the machinery_opensuse_<system>_spec.rb files and
pass it to include_examples_for_platform.

- Tests are selected from integration_tests.yml
include_example_for_platform reads the integration test definition
from files and parse it for tests that are marked as "full_test"
"acceptance_tests" or "minimal_tests" for the given system. It compiles
a array with tests and then runs the corresponding list of test/system
combinations.

- Test groups to be run are selected via environment variable
The tests to be executed can be selected via the shell environment
variable TESTGROUP that needs to be set to "full", "acceptance" or
"minimal". If the variable is empty, the full test set is executed to
keep the default behaviour of the previous test execution method.
Test groups are inclusive, full tests include acceptance and minimal
tests and minimal tests include acceptance tests.

- Test descriptions are changed
Test descriptions now are used to determine the machinery host and
therefor are changed to include the system name as it is used in the
yaml file.

- Vagrant names for machinery hosts are adjusted
The names for machinery hosts instances in the Vagrantfile are adjusted
to the names used in the yaml file. We prefix them with "machinery_" to
distinguish them from the test systems that the tests are run against.

- Vagrant names for test hosts get mapped
The vagrant names for the system the test are run against are mapped to
the names used in the yaml file.