Provide export integration test descriptions for multiple systems.
- Add required descriptions for supported systems
Exported tests now can be run for different systems. This commit
provides descriptions for supported systems.

- Descriptions based on existing opensuse13.1 description
The openSUSE_13_1 files are a copy of the already existing files in
spec/data/descriptions/jeos. The file sections of the other added
descriptions are also a copy of the existing description while the
other scopes are unique for the different supported systems.

- Add expected files for export tests
Add corresponding expected files for autoyast and kiwi export tests