Add first version of the Environment scope
This scope will be used to store information about the inspected system.
For now it will contain the locale that's being used for executing
commands by machinery.

This scope will be made hidden in a later commit because it's a
prerequisite for executing commands and thus has to be run on each
inspection. Making it visible would create some potential for confusion
or breakage (e.g. users trying to exclude the scope etc.)