Import machinery-helper sources to machinery
The CHANGES.md, machinery-helper.spec.erb, machinery-helper-rpmlintrc
and RPM_CHANGES files and the the Makefile were removed because they
are now obsolete.

Additionally the license header was added.

The version was also increased to 0.0.4 but will be replaced with the
following commits anyway.

Original Changelog:
* Add --version parameter to show the version
* Add support for packaging files and directories as tar.gz and piping
* the content through stdout
*  Add rpc_pipefs to filtered filesystems
* Implemented inspection of unmanaged-files

Remove GO Makefile

becaues it is obsolete now with the Rakefile.