Drop unnecessary `--docker` option for `inspect-container` command
Right now we only support docker containers for the `inspect-container`
command.
Because of that the logic behind the `--docker` option was already used
as default.
That is the reason for removing the option for now.
We will look for solutions how to define the container type for the
command properly
as soon as we support more types.