Fix Inspector Files script
The inspector files script was created to easily gather the test data
for the integration test of new systems.
It was never used before and had quite some issues. So files weren't
extracted although our integration tests always run the exctraction.
Then the wrong class was used in the Rakefile, the name was most likely
changed during the review.
And Hound complained about gsub so it was replaced by tr.