Run standalone tests if target system is local in matrix
We still have tests like validate or CLI that are not triggered via our
test matrix, so these tests are run in every test run. As we want to use
the test matrix also for prophet tests, we need to run those via matrix
to avoid them being run for all system on every prophet run.
This commit executes tests when they are defined to be run locally on the
machinery host.