Clean up HTTP server code and its messages
It should be important that warnings should be understandable and should
inform users about the risks by sharing descriptions over the network.

The usage of the class has changed a bit. It's now possible to set
either an IP address of the HTTP server or the `public` flag. The public
flag makes the server listen on all configured IP addresses.