Extend yum repositories format with missing options
In the past we only used the elements of yum which matched with our
zypper model.
Since we had to separate the repositories now anyway because of the
apt support we were able to add common elements.

In this succession the attributes gpgkey and mirrorlist where added.
Additionally instead of just using the first array element of url
we now support arrays for them.

So url and gpgkey are arrays.