Split apt-cache output to prevent false matches
The Debian package matcher for extracting the MD5sum and Vendor didn't
work correctly because it matched more than one package at a time.

This is now prevented by splitting each package up and only match
against these parts.

It is also possible to sometimes have an Origin but no MD5sum, that's
why the matching is split up for both.