Fix tumbleweed network device name issues
Originally the tumbleweed images had device names eth0 and eth1 and all
went well with these names. Since 20160207 this changed with the new
tumbleweed build and devices got renamed to ens5 and ens6. Our initial fix
was to just configure these interfaces for the new names, but that failed
to take into account that vagrant wants eth0 and especially eth1 to setup
its management network. After some research it became clear that the whole
renaming can be avoided in the first place by telling systemd via grub
commandline to not do that and leave the device names with the names
originally given to them by the kernel - eth0 and eth1.