Use static bootstrap list for Kiwi export
As it turns out Kiwi only needs a very basic list of bootstrap
packages which are not supposed to change betweend distros so
there is no point for us to put all packages under bootstrap.

The package "filesystem" is mandatory and everything is only put to
bootstrap if it exists in the system description.

Kiwi config