Fix flaky `machinery serve` integration test
The problem was caused by the helper compilation step. The serve test is
the first test that is run locally and thus triggers the compilation
when necessary. That step takes quite a while, though, and thus causes
the 10 second timeout in the serve test to break.

This problem is now fixed by always making sure that the helper is
compiled before the actual test suite is run.