Replace click with JS based trigger click to close modal dialog in specs
Using click the feature spec fails about one in ten times. Pressing
the button to close the modal dialog triggers a fade animation. Due
to some timing issue poltergeist does not always close the dialog and
the spec fails.
While trigger("Click") is different from a real users action it
reliably closes the modal dialog.

Alternatively it may be possible to disable CSS animations altogether
while testing. However this is not easy to implement, as it would involve
injecting JS for tests, or manipulating CSS classes, without affecting
the functionality of the dialog.