Read files on a remote system without sudo by default
If read_file was used on a remote system, sudo was used all
the time.
To prevent this the optional option :privileged was added.

The default was switched to non-privileged because there was only one
file which needed these rights.

Local and Docker System classes do not use sudo because they
already have root rights.