Do not extract Metadata for missing unmanaged files
Make sure that files exist before extracting the metadata and remove
them from the unmanaged-files list if not.

This should allow filtering to work properly because the filters are
applied before extraction so it is possible for the user to filter
directories with changing files.
Before that machinery-helper could crash if a file was removed during
the first phase of inspection and there was no way for the user to
filter these directories because the helper does not support filtering
yet.