Fix build of documentation
by providing support for newer mkdocs versions.

With mkdocs versions >= 1.0 many of the Markdown links did not work
anymore so the directory structure had to be changed to fix these issue.
These changes also make the directory structure more consistent.

While doing so I have also removed a couple of duplicate files because
they can be now properly referenced.

Besides that other config options needed to be changed to be compatible
with mkdocs >= 1.0.